refactor: move data treating functions to separate file
